# lab3/db-server/go/api_execution.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * MIMUW ISBD database system
 *
 * This file describes interface between DBMS system and user.
 *
 * API version: 1.0.1
 */

package openapi

import (
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// ExecutionAPIController binds http requests to an api service and writes the service results to the http response
type ExecutionAPIController struct {
	service      ExecutionAPIServicer
	errorHandler ErrorHandler
}

// ExecutionAPIOption for how the controller is set up.
type ExecutionAPIOption func(*ExecutionAPIController)

// WithExecutionAPIErrorHandler inject ErrorHandler into controller
func WithExecutionAPIErrorHandler(h ErrorHandler) ExecutionAPIOption {
	return func(c *ExecutionAPIController) {
		c.errorHandler = h
	}
}

// NewExecutionAPIController creates a default api controller
func NewExecutionAPIController(s ExecutionAPIServicer, opts ...ExecutionAPIOption) *ExecutionAPIController {
	controller := &ExecutionAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the ExecutionAPIController
func (c *ExecutionAPIController) Routes() Routes {
	return Routes{
		"GetQueries": Route{
			"GetQueries",
			strings.ToUpper("Get"),
			"/queries",
			c.GetQueries,
		},
		"GetQueryById": Route{
			"GetQueryById",
			strings.ToUpper("Get"),
			"/query/{queryId}",
			c.GetQueryById,
		},
		"SubmitQuery": Route{
			"SubmitQuery",
			strings.ToUpper("Post"),
			"/query",
			c.SubmitQuery,
		},
		"GetQueryResult": Route{
			"GetQueryResult",
			strings.ToUpper("Get"),
			"/result/{queryId}",
			c.GetQueryResult,
		},
		"GetQueryError": Route{
			"GetQueryError",
			strings.ToUpper("Get"),
			"/error/{queryId}",
			c.GetQueryError,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the ExecutionAPIController
func (c *ExecutionAPIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"GetQueries",
			strings.ToUpper("Get"),
			"/queries",
			c.GetQueries,
		},
		Route{
			"GetQueryById",
			strings.ToUpper("Get"),
			"/query/{queryId}",
			c.GetQueryById,
		},
		Route{
			"SubmitQuery",
			strings.ToUpper("Post"),
			"/query",
			c.SubmitQuery,
		},
		Route{
			"GetQueryResult",
			strings.ToUpper("Get"),
			"/result/{queryId}",
			c.GetQueryResult,
		},
		Route{
			"GetQueryError",
			strings.ToUpper("Get"),
			"/error/{queryId}",
			c.GetQueryError,
		},
	}
}

// GetQueries - Get list of queries (optional in project 3, but useful). Use those IDs to get details by calling /query endpoint.
func (c *ExecutionAPIController) GetQueries(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetQueries(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetQueryById - Get detailed status of selected query
func (c *ExecutionAPIController) GetQueryById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	queryIdParam := params["queryId"]
	if queryIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"queryId"}, nil)
		return
	}
	result, err := c.service.GetQueryById(r.Context(), queryIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// SubmitQuery - Submit new query for execution
func (c *ExecutionAPIController) SubmitQuery(w http.ResponseWriter, r *http.Request) {
	var executeQueryRequestParam ExecuteQueryRequest
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&executeQueryRequestParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertExecuteQueryRequestRequired(executeQueryRequestParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertExecuteQueryRequestConstraints(executeQueryRequestParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.SubmitQuery(r.Context(), executeQueryRequestParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetQueryResult - Get result of selected query (will be available only for SELECT queries after they are completed)
func (c *ExecutionAPIController) GetQueryResult(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	queryIdParam := params["queryId"]
	if queryIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"queryId"}, nil)
		return
	}
	var getQueryResultRequestParam GetQueryResultRequest
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&getQueryResultRequestParam); err != nil && !errors.Is(err, io.EOF) {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertGetQueryResultRequestRequired(getQueryResultRequestParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertGetQueryResultRequestConstraints(getQueryResultRequestParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.GetQueryResult(r.Context(), queryIdParam, getQueryResultRequestParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetQueryError - Get error of selected query (will be available only for queries in FAILED state)
func (c *ExecutionAPIController) GetQueryError(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	queryIdParam := params["queryId"]
	if queryIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"queryId"}, nil)
		return
	}
	result, err := c.service.GetQueryError(r.Context(), queryIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}