# lab3/db-server/go/api_proj3.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * MIMUW ISBD database system
 *
 * This file describes interface between DBMS system and user.
 *
 * API version: 1.0.1
 */

package openapi

import (
	"encoding/json"
	"errors"
	"io"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// Proj3APIController binds http requests to an api service and writes the service results to the http response
type Proj3APIController struct {
	service      Proj3APIServicer
	errorHandler ErrorHandler
}

// Proj3APIOption for how the controller is set up.
type Proj3APIOption func(*Proj3APIController)

// WithProj3APIErrorHandler inject ErrorHandler into controller
func WithProj3APIErrorHandler(h ErrorHandler) Proj3APIOption {
	return func(c *Proj3APIController) {
		c.errorHandler = h
	}
}

// NewProj3APIController creates a default api controller
func NewProj3APIController(s Proj3APIServicer, opts ...Proj3APIOption) *Proj3APIController {
	controller := &Proj3APIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the Proj3APIController
func (c *Proj3APIController) Routes() Routes {
	return Routes{
		"GetTables": Route{
			"GetTables",
			strings.ToUpper("Get"),
			"/tables",
			c.GetTables,
		},
		"GetTableById": Route{
			"GetTableById",
			strings.ToUpper("Get"),
			"/table/{tableId}",
			c.GetTableById,
		},
		"DeleteTable": Route{
			"DeleteTable",
			strings.ToUpper("Delete"),
			"/table/{tableId}",
			c.DeleteTable,
		},
		"CreateTable": Route{
			"CreateTable",
			strings.ToUpper("Put"),
			"/table",
			c.CreateTable,
		},
		"GetQueries": Route{
			"GetQueries",
			strings.ToUpper("Get"),
			"/queries",
			c.GetQueries,
		},
		"GetQueryById": Route{
			"GetQueryById",
			strings.ToUpper("Get"),
			"/query/{queryId}",
			c.GetQueryById,
		},
		"SubmitQuery": Route{
			"SubmitQuery",
			strings.ToUpper("Post"),
			"/query",
			c.SubmitQuery,
		},
		"GetQueryResult": Route{
			"GetQueryResult",
			strings.ToUpper("Get"),
			"/result/{queryId}",
			c.GetQueryResult,
		},
		"GetQueryError": Route{
			"GetQueryError",
			strings.ToUpper("Get"),
			"/error/{queryId}",
			c.GetQueryError,
		},
		"GetSystemInfo": Route{
			"GetSystemInfo",
			strings.ToUpper("Get"),
			"/system/info",
			c.GetSystemInfo,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the Proj3APIController
func (c *Proj3APIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"GetTables",
			strings.ToUpper("Get"),
			"/tables",
			c.GetTables,
		},
		Route{
			"GetTableById",
			strings.ToUpper("Get"),
			"/table/{tableId}",
			c.GetTableById,
		},
		Route{
			"DeleteTable",
			strings.ToUpper("Delete"),
			"/table/{tableId}",
			c.DeleteTable,
		},
		Route{
			"CreateTable",
			strings.ToUpper("Put"),
			"/table",
			c.CreateTable,
		},
		Route{
			"GetQueries",
			strings.ToUpper("Get"),
			"/queries",
			c.GetQueries,
		},
		Route{
			"GetQueryById",
			strings.ToUpper("Get"),
			"/query/{queryId}",
			c.GetQueryById,
		},
		Route{
			"SubmitQuery",
			strings.ToUpper("Post"),
			"/query",
			c.SubmitQuery,
		},
		Route{
			"GetQueryResult",
			strings.ToUpper("Get"),
			"/result/{queryId}",
			c.GetQueryResult,
		},
		Route{
			"GetQueryError",
			strings.ToUpper("Get"),
			"/error/{queryId}",
			c.GetQueryError,
		},
		Route{
			"GetSystemInfo",
			strings.ToUpper("Get"),
			"/system/info",
			c.GetSystemInfo,
		},
	}
}

// GetTables - Get list of tables with their accompanying IDs. Use those IDs to get details by calling /table endpoint.
func (c *Proj3APIController) GetTables(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetTables(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetTableById - Get detailed description of selected table
func (c *Proj3APIController) GetTableById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	tableIdParam := params["tableId"]
	if tableIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"tableId"}, nil)
		return
	}
	result, err := c.service.GetTableById(r.Context(), tableIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteTable - Delete selected table from database
func (c *Proj3APIController) DeleteTable(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	tableIdParam := params["tableId"]
	if tableIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"tableId"}, nil)
		return
	}
	result, err := c.service.DeleteTable(r.Context(), tableIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// CreateTable - Create new table in database
func (c *Proj3APIController) CreateTable(w http.ResponseWriter, r *http.Request) {
	var tableSchemaParam TableSchema
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&tableSchemaParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertTableSchemaRequired(tableSchemaParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertTableSchemaConstraints(tableSchemaParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.CreateTable(r.Context(), tableSchemaParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetQueries - Get list of queries (optional in project 3, but useful). Use those IDs to get details by calling /query endpoint.
func (c *Proj3APIController) GetQueries(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetQueries(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetQueryById - Get detailed status of selected query
func (c *Proj3APIController) GetQueryById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	queryIdParam := params["queryId"]
	if queryIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"queryId"}, nil)
		return
	}
	result, err := c.service.GetQueryById(r.Context(), queryIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// SubmitQuery - Submit new query for execution
func (c *Proj3APIController) SubmitQuery(w http.ResponseWriter, r *http.Request) {
	var executeQueryRequestParam ExecuteQueryRequest
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&executeQueryRequestParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertExecuteQueryRequestRequired(executeQueryRequestParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertExecuteQueryRequestConstraints(executeQueryRequestParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.SubmitQuery(r.Context(), executeQueryRequestParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetQueryResult - Get result of selected query (will be available only for SELECT queries after they are completed)
func (c *Proj3APIController) GetQueryResult(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	queryIdParam := params["queryId"]
	if queryIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"queryId"}, nil)
		return
	}
	var getQueryResultRequestParam GetQueryResultRequest
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&getQueryResultRequestParam); err != nil && !errors.Is(err, io.EOF) {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertGetQueryResultRequestRequired(getQueryResultRequestParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertGetQueryResultRequestConstraints(getQueryResultRequestParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.GetQueryResult(r.Context(), queryIdParam, getQueryResultRequestParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetQueryError - Get error of selected query (will be available only for queries in FAILED state)
func (c *Proj3APIController) GetQueryError(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	queryIdParam := params["queryId"]
	if queryIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"queryId"}, nil)
		return
	}
	result, err := c.service.GetQueryError(r.Context(), queryIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetSystemInfo - Get basic information about the system (e.g. version, uptime, etc.)
func (c *Proj3APIController) GetSystemInfo(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetSystemInfo(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}